requirements.txt
Flask==2.0.1
tensorflow==2.5.0
numpy==1.19.5
Pillow==8.2.0
